pyproject.toml
[project]
name = "easy-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiolimiter>=1.1.0",
    "black>=24.10.0",
    "polars>=1.9.0",
    "python-dotenv>=1.0.1",
    "reflex-ag-grid>=0.0.8",
    "reflex>=0.6.0",
    "xlsxwriter>=3.2.0",
]
